feat(settings): SettingsTabs + combined /settings/jobs (#227)
Ship the tracer-bullet slice of the Settings redesign (PRD #226):

- SettingsTabs: URL-synced (?tab=<key>) tabbed wrapper around the
  existing @base-ui/react/tabs primitive. Reads the active tab from the
  query string, falls back to defaultTab then first tab; updates the URL
  via router.replace on switch.
- /settings/jobs: first combined section. Three tabs — Statuses, Damage
  Types, Intake Form — render the three existing page bodies, extracted
  into co-located *-tab.tsx components (content unchanged).
- Server-side redirects from /settings/{statuses,damage-types,intake-form}
  to /settings/jobs?tab=<key> so bookmarks keep working.
- settings-nav.ts: one new Jobs entry, three old entries removed,
  Navigation entry removed (slated for a main-nav reorder mode).
- CONTEXT.md: add "Outgoing email" domain term used by slices 4 and 8.
