+++ src/lib/settings-nav.ts
@@ -2,18 +2,15 @@ import {
   Building2,
   Palette,
   ListChecks,
-  Flame,
   Store,
   Receipt,
   Users,
   Mail,
   FileSignature,
-  ClipboardList,
   Bell,
   FileText,
   Download,
   BookOpen,
-  Menu,
   Send,
   Link2,
   CreditCard,
@@ -32,9 +29,7 @@ export interface SettingsNavItem {
 export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/company", label: "Company Profile", icon: Building2 },
   { href: "/settings/appearance", label: "Appearance", icon: Palette },
-  { href: "/settings/navigation", label: "Navigation", icon: Menu },
-  { href: "/settings/statuses", label: "Job Statuses", icon: ListChecks },
-  { href: "/settings/damage-types", label: "Damage Types", icon: Flame },
+  { href: "/settings/jobs", label: "Jobs", icon: ListChecks },
   { href: "/settings/item-library", label: "Item Library", icon: Library },
   { href: "/settings/estimate-templates", label: "Estimate Templates", icon: LayoutTemplate },
   { href: "/settings/pdf-presets", label: "PDF Presets", icon: FileText },
@@ -43,7 +38,6 @@ export const settingsNavItems: SettingsNavItem[] = [
   { href: "/settings/users", label: "Users & Crew", icon: Users },
   { href: "/settings/email", label: "Email Accounts", icon: Mail },
   { href: "/settings/signatures", label: "Email Signatures", icon: FileSignature },
-  { href: "/settings/intake-form", label: "Intake Form", icon: ClipboardList },
   { href: "/settings/contract-templates", label: "Contract Templates", icon: FileText },
   { href: "/settings/contracts", label: "Contracts", icon: Send },
   { href: "/settings/notifications", label: "Notifications", icon: Bell },
